// server/types.ts
export type UserRole = 'siswa' | 'guru_bk';

export interface User {
  id: string;
  name: string;
  username: string;
  password_hash: string;
  role: UserRole;
  nis?: string;
  class?: string;
  major?: string;
  avatar_url?: string;
  must_change_password?: boolean;
  created_at: string;
  updated_at: string;
}